Use usize_t for loop index

+++ src/debug.c
@@ -10,7 +10,7 @@ static int constantLongInstruction(const char *name, Chunk *chunk, int offset);
 
 void disassembleChunk(Chunk *chunk, const char *name) {
   printf("== %s ==\n", name);
-  for (int offset = 0; offset < chunk->count;) {
+  for (size_t offset = 0; offset < chunk->count;) {
     offset = disassembleInstruction(chunk, offset);
   }
 }
